Split out guts of insertAutoCopyAutoDestroy in to OwnershipFlowManager.
The interfaces to functions manipulating the flow variables were getting unwieldly, so I
wrapped all of that state information in the OwnershipFlowManager class.  I also split out
the AliasVectorMap into a separate header file.

This is just a code reorganization with no functional changes.
